chore: add mvvm layer to settings page
